♻️ Relocates shared pydantic to new shared directory directory
Moves pydantic related shared code from a dedicated 'pydanticde' package within the pydantic layer to a new '_pydantic' directory which is added like _shared.

This change organises the project structure, placing pydantic models alongside the functions that use them, rather than in a dedicated layer.
