--- functions/_layers/pydantic/python/pydanticde/__init__.py
@@ -1,6 +1,0 @@
-"""
-pydanticde package.
-
-This package provides a collection of models and functions which are related to the Dance Engine implementation of pydantic.
-"""
-
