+++ functions/_pydantic/EventBridge.py
