+++ functions/functions/customers/sls.customers.function.yml
@@ -7,6 +7,7 @@ Customers:
       - '!**/**'
       - "functions/customers/**"
       - "_shared/**"
+      - "_pydantic/**"
   environment:
     CORE_TABLE_NAME: ${self:custom.core_table_name}
     ORG_TABLE_NAME_TEMPLATE: ${self:custom.dynamodb_table_format}
