+++ functions/functions/events/sls.events.function.yml
@@ -7,6 +7,7 @@ Events:
         - '!**/**'
         - "functions/events/**"
         - "_shared/**"
+        - "_pydantic/**"      
     environment:
       ORG_TABLE_NAME_TEMPLATE: ${self:custom.dynamodb_table_format}
     layers:
