+++ functions/functions/utils/history_tracker/lambda_history_tracker.py
@@ -3,8 +3,8 @@ import json
 import boto3 
 import logging
 
-from pydanticde.dynamodb import HistoryModel, convert_floats_to_decimals # pydantic layer
-from pydanticde.EventBridge import EventBridgeEvent, Action, EventType, EventBridgeEventDetail # pydantic layer
+from _pydantic.dynamodb import HistoryModel, convert_floats_to_decimals # pydantic layer
+from _pydantic.EventBridge import EventBridgeEvent, Action, EventType, EventBridgeEventDetail # pydantic layer
 
 from _shared.DecimalEncoder import DecimalEncoder
